IBE implement based on JPBC and Java JCA framework
This project is an implementation of IBE based the JPBC library. If
obtained the JCE certificate, this IBE implementation could be used in
the same way as other encryption algorithms in the JCA framework. The
project ChiralBehaviors/IBE is very helpful to me. Parts of the codes
in the key management are from the book Java Cryptography by Jonathan
B. Knudsen. Although published almost twenty years ago, the book is a
great reference book for implementing and applying cryptography.